Move gcpu to proper directory so unit-test manifest works
